style: 좋아요 및 북마크 버튼에 반응형 적용 (#249)
