+++ src/app/_components/client/home-popular-schedules.tsx
@@ -22,7 +22,7 @@ const PopularSchedules = () => {
 
   // 스켈레톤 UI 렌더링
   const renderSkeleton = () => (
-    <div className="grid w-full grid-cols-1 gap-x-3 gap-y-5 sm:grid-cols-2 md:grid-cols-3">
+    <div className="grid w-full list-none grid-cols-2 gap-x-3 gap-y-5 p-0 md:grid-cols-3">
       {Array(6)
         .fill(null)
         .map((_, index) => (
@@ -35,7 +35,7 @@ const PopularSchedules = () => {
   );
 
   return (
-    <div className="flex w-full max-w-[1024px] flex-col gap-[17.5px] p-9">
+    <div className="flex w-full max-w-[1024px] flex-col gap-[17.5px] p-4 md:p-7 lg:p-9">
       <div className="flex items-center justify-between">
         <h2 className="semibold-22 text-center text-gray-900">
           올레 인기 일정
@@ -51,7 +51,7 @@ const PopularSchedules = () => {
       {isLoading ? (
         renderSkeleton()
       ) : (
-        <div className="grid w-full grid-cols-1 gap-x-3 gap-y-5 sm:grid-cols-2 md:grid-cols-3">
+        <div className="grid w-full list-none grid-cols-2 gap-x-3 gap-y-5 p-0 md:grid-cols-3">
           {plans.map((plan: PlanType) => (
             <PlanVerticalCard key={plan.planId} plan={plan} />
           ))}
