+++ src/app/categories/_components/_client/category-client.tsx
@@ -57,7 +57,7 @@ const CategoryClient = ({ category }: { category: CategoryParamType }) => {
 
       {/* 카드 그리드 */}
       <section>
-        <ul className="grid list-none grid-cols-2 gap-[11px] p-0 sm:grid-cols-3 md:grid-cols-4">
+        <ul className="grid list-none grid-cols-2 gap-x-[11px] gap-y-[22px] p-0 sm:grid-cols-3 md:grid-cols-4">
           {isPending ? (
             <PlaceListSkeleton count={16} />
           ) : (
@@ -76,7 +76,7 @@ const CategoryClient = ({ category }: { category: CategoryParamType }) => {
                 {index === 7 && (
                   <li
                     key={`banner-${index}`}
-                    className="col-span-full my-4 flex w-full items-center justify-center"
+                    className="col-span-full my-2 flex w-full items-center justify-center"
                   >
                     <Banner />
                   </li>
