+++ src/components/features/card/plan-vertical-card.tsx
@@ -4,8 +4,7 @@ import Duration from '@/components/commons/duration';
 import PlanImage from '@/components/commons/plan-image';
 import { Separator } from '@/components/ui/separator';
 import { PlanVerticalCardProps } from '@/types/plan.type';
-
-import LikeButton from '../like/like-button';
+import LikeButton from '@/components/commons/like-button';
 
 /**
  * vertical한 플랜 카드 컴포넌트
